% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdownsampling.R
\name{gdownsample_mat}
\alias{gdownsample_mat}
\title{Gene Expression Down-sampling for Matrix}
\usage{
gdownsample_mat(mat, mean, sd)
}
\arguments{
\item{mat}{The input matrix that will be down-sampled}

\item{mean}{The mean parameter for down-sampling}

\item{sd}{The standard deviation parameter for down-sampling}
}
\value{
The matrix after down-sampling
}
\description{
Down-sample gene expression based on matrix
}
\examples{
library(scRNAseq)
sce = ReprocessedAllenData("tophat_counts")
counts = assay(sce, 'tophat_counts')
gdownsample_mat(counts, mean=0.3, sd=0.05)
}
