% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdownsampling.R
\name{gdownsample_obj}
\alias{gdownsample_obj}
\title{Gene Expression Down-sampling for Seurat or SingleCellExperiment Object}
\usage{
gdownsample_obj(object, user_assay = NULL, returnMode = "new", mean, sd)
}
\arguments{
\item{object}{The input Seurat or SingleCellExperiment object for down-sampling}

\item{user_assay}{The input assay for down-sampling. For Seurat object, if not specified, down-sampling will be performed based on the active assay. For SingleCellExperiment object, an assay has to be specified.}

\item{returnMode}{The return mode for down-sampling. Default is "new": creating a new down-sampled object while keep the original object. Other two available options are 1. "add": add a new down-sampled assay. 2. "replace": add a new down-sampled assay and delete the original assay.}

\item{mean}{The mean parameter for down-sampling}

\item{sd}{The standard deviation parameter for down-sampling}
}
\value{
A Seurat/SingleCellExperiment object or none based on the return mode parameter.
}
\description{
Down-sample gene expression based on Seurat or SingleCellExperiment object
}
\examples{
library(scRNAseq)
sce = ReprocessedAllenData("tophat_counts")
counts = assay(sce, 'tophat_counts')
gdownsample_mat(counts, mean=0.3, sd=0.05)
}
